// src/validate/types.ts
export type ModelValidationAnnotationCache = Map<
  string,
  Map<AnnotateDefinition, AnnotationData>
>;

export type ValidateDefinitionErrors<ErrorType> = Map<
  ValidateDefinition<ErrorType>,
  { field: string; errors: ErrorType[] }
>;

export type MutatingErrorCacheKey<ErrorType> =
  | RecursiveDefinition<ErrorType>
  | number;

export type ModelValidationCache<ErrorType> = {
  isPathActive: boolean;
  annotations: ModelValidationAnnotationCache;
  errors: ValidateDefinitionErrors<ErrorType>;
  children: Map<
    MutatingErrorCacheKey<ErrorType>,
    ModelValidationCache<ErrorType>
  >;
};

// { [field]: { [annotation]: Set<subscriptions> }}
export type FieldAnnotationSubscriptions = Map<
  string,
  Map<Annotation<any>, Set<FieldAnnotationSubscription<any, any>>>
>;

export type AnnotationsSubscriptions = Set<AnnotationsSubscription>;

// the Map keys are paths, and the values are set of changed annotations
// we use Map & Set here so we don't store duplicate entries to make processing faster
export type ChangedAnnotationsCache = Map<string, Set<Annotation<any>>>;

export type ResolvedValidations<ErrorType> = {
  current: Record<string, ErrorType[]>;
};

export type ResolvedAnnotations = {
  current: Record<string, AnnotationData>;
};

export type ModelValidationContext<
  Data,
  ExternalData = undefined,
  ErrorType = string
> = {
  model: Model<Data, ExternalData, ErrorType>;
  previousData: Data | undefined;
  previousExternalData: ExternalData | undefined;
  cache: ModelValidationCache<ErrorType>;
  resolvedAnnotations: ResolvedAnnotations;
  resolvedValidations: ResolvedValidations<ErrorType>;
  subscriptions: {
    fieldAnnotation: FieldAnnotationSubscriptions;
    annotations: AnnotationsSubscriptions;
  };
  pendingChangedAnnotations: ChangedAnnotationsCache;
  transactionCounter: number;
};